meeting-manager-agent/requirements.txt
litellm
todoist_api_python 
colorama
python-dotenv
